Refactor resolve to remove double callbacks
Resolve previously offered two callbacks for methods that saved
proposals which updated events and search indexes and notifications, as
an optimization: clients would receive a response on the socket for
updates before the extra time taken to talk to solr and mongo to update
these indexes.

This commit removes that in order to simplify at the expense of the
performance optimization.