Add resolution history with messages for proposals
Adds a new array to the resolve proposal schema which tracks the date
and details of resolutions and reopenings of proposals. This makes it
possible to include finalizing and re-opening of proposals in the
history view for a proposal.

This ought to gracefully degrade to older models which do not have this
data stored.

The resolution fields duplicate the bare "resolved" and "passed" fields in
the proposal model; for now I'm leaving them in for simplicity of access
and to avoid requiring a migration.