Add 'name' field to twinklepad, doc-up tp front
Instead of allowing users to change the "pad_name" parameter, which
leads to issues of syncing with the immutable "padName" for etherpads,
use the "pad_name" as a set-once-only slug, and add an additional "name"
field which is mutable.  Restore the "resync_etherpad_ids.js" script
which is now correct again.

Also, add documentation for how to use twinklepad to its front page,
rather than the lists of documents that were there previously.  Include
a link to the dashboard to list existing pads.

Add a "/dashboard/" route that redirects to login if not signed in, and
then redirects to "/".  This makes for a robust link target that will
result in the user being signed in at their dashboard, unlike "/" which
would show the splash page if not signed in, or the login page which
would confuse if the user is already logged in.