Accord dates in email and web for activity summary
The web page uses a URL with /YYYY/mm/dd.  It interprets that as "from
12:00:00 the morning of 11:59:59 the evening of YYYY-mm-dd".  The email,
though, wants to send a comprehensive list of activity, but does not
send at midnight necessarily.

Make the email span from (now() - 24 hours).  Interpret "now" to be the
YYYY/mm/dd date for the purpose of linking.  This means that the web
will differ from email content for anything between the hours of when
the email is sent and the end of the day -- the email will show more,
extending past the previous midnight.  But this is arguably less obvious
anyway; so let's just rely on the "previous day" / "next day" links for
that.