Don't send activity summary if it's only invites
When someone is invited, they are sent an email. This should be the last
email they receive until they act on the invitation and have other
activity.  Opportunistically include invitations if someone would
receive an activity summary anyway; but filter out those that are just
invitations.

We don't want to be linked in style spammers.