Refactor authentication and persona watch
Previously, we did not specify loggedInUser in the call to
navigator.id.watch, resulting in persona calling "onlogin" on every page
load for authenticated users.  Ref:
https://developer.mozilla.org/en-US/docs/Web/API/navigator.id.watch
Consequently, every page load would result in the backend verifying an
assertion, and refreshing the session, and returning the results to the
client.

We want to preserve the session refresh, but do away with the needless
extra assertion verification. So we now specify loggedInUser, but add a
new socket request on page load which calls "refresh_session".  This
call retrieves all current users and groups, stores them to the
server-side session, and re-notifies rooms of presence.