Switch to polling to confer hangout room name
The frame that google uses to load our app has a random origin --
something like:
https://a6ifkkk3qdf3vgqrl47pfiikqjcmllng-a-hangout-opensocial.googleusercontent.com.

Since we can't be sure what this will be, we can't post a message to it
from the interior frame.  The exterior frame can't know when all the
scripts on the interior are finished loading and ready to receive the
room name.

Solution: make the exterior frame blast the interior with a message
informing it of the hangout URL. When the interior frame gets it, it
responds using the origin of the sender with an "ack", telling the
exterior frame to stop blasting.

Best security practices for cross-document messaging require careful
checking of the origin. Instead of defining config.hangout_origin which
we match exactly for messages to the interior frame, set a regex in
config.hangout_origin_re.  That way we can match for e.g.:

^https://[-a-z0-9]+hangout-opensocial.googleusercontent.com$